move the menu implementations into the core module
I think these are a good candidate to become their own module but that's not something I'm going to work on today.
