+++ agonyforge-mud-core/src/main/java/com/agonyforge/mud/core/cli/menu/impl/MenuPrompt.java
@@ -1,4 +1,4 @@
-package com.agonyforge.mud.demo.cli.menu;
+package com.agonyforge.mud.core.cli.menu.impl;
 
 import com.agonyforge.mud.core.cli.menu.AbstractMenuPrompt;
 import com.agonyforge.mud.core.cli.Color;
