+++ agonyforge-mud-core/src/main/java/com/agonyforge/mud/core/cli/menu/impl/MenuTitle.java
@@ -1,4 +1,4 @@
-package com.agonyforge.mud.demo.cli.menu;
+package com.agonyforge.mud.core.cli.menu.impl;
 
 import com.agonyforge.mud.core.cli.menu.AbstractMenuTitle;
 import com.agonyforge.mud.core.cli.Color;
