+++ agonyforge-mud-core/src/test/java/com/agonyforge/mud/core/cli/menu/impl/MenuItemTest.java
@@ -1,7 +1,9 @@
-package com.agonyforge.mud.demo.cli.menu;
+package com.agonyforge.mud.core.cli.menu.impl;
 
 import com.agonyforge.mud.core.cli.Color;
+import com.agonyforge.mud.core.cli.menu.impl.MenuItem;
 import com.agonyforge.mud.core.web.model.Output;
+import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
@@ -15,26 +17,30 @@ import static com.agonyforge.mud.core.cli.Color.RED;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 @ExtendWith(MockitoExtension.class)
-public class MenuTitleTest {
+public class MenuItemTest {
+    @Test
+    void testGetItem() {
+        String foo = "foo";
+        MenuItem uut = new MenuItem("T", "Test", foo);
+
+        assertEquals(foo, uut.getItem());
+    }
+
     @ParameterizedTest
     @MethodSource
     void testRender(Color[] colors, String expected) {
-        MenuTitle uut = new MenuTitle("Test");
+        MenuItem uut = new MenuItem("T", "Test");
         Output result = uut.render(colors);
-        String dividerText = (colors.length > 1 ? "[red]" : "[green]") + "*".repeat(10);
 
-        assertEquals(4, result.getOutput().size());
-        assertEquals("", result.getOutput().get(0));
-        assertEquals(dividerText, result.getOutput().get(1));
-        assertEquals(expected, result.getOutput().get(2));
-        assertEquals(dividerText, result.getOutput().get(3));
+        assertEquals(1, result.getOutput().size());
+        assertEquals(expected, result.getOutput().get(0));
     }
 
     private static Stream<Arguments> testRender() {
         return Stream.of(
-            Arguments.of(new Color[] {BLUE, RED}, "[red]* [blue] Test [red] *"),
-            Arguments.of(new Color[] {BLUE}, "[green]* [blue] Test [green] *"),
-            Arguments.of(new Color[] {}, "[green]* [yellow] Test [green] *")
+            Arguments.of(new Color[] {BLUE, RED}, "[blue]T[red]) [blue]Test"),
+            Arguments.of(new Color[] {BLUE}, "[blue]T[green]) [blue]Test"),
+            Arguments.of(new Color[] {}, "[yellow]T[green]) [yellow]Test")
         );
     }
 }
