+++ agonyforge-mud-core/src/test/java/com/agonyforge/mud/core/cli/menu/impl/MenuPromptTest.java
@@ -1,5 +1,6 @@
-package com.agonyforge.mud.demo.cli.menu;
+package com.agonyforge.mud.core.cli.menu.impl;
 
+import com.agonyforge.mud.core.cli.menu.impl.MenuPrompt;
 import com.agonyforge.mud.core.web.model.Output;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
