+++ agonyforge-mud-core/src/test/java/com/agonyforge/mud/core/cli/menu/impl/MenuTitleTest.java
@@ -1,6 +1,7 @@
-package com.agonyforge.mud.demo.cli.menu;
+package com.agonyforge.mud.core.cli.menu.impl;
 
 import com.agonyforge.mud.core.cli.Color;
+import com.agonyforge.mud.core.cli.menu.impl.MenuTitle;
 import com.agonyforge.mud.core.web.model.Output;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -15,22 +16,26 @@ import static com.agonyforge.mud.core.cli.Color.RED;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 @ExtendWith(MockitoExtension.class)
-public class MenuDividerTest {
+public class MenuTitleTest {
     @ParameterizedTest
     @MethodSource
     void testRender(Color[] colors, String expected) {
-        MenuDivider uut = new MenuDivider(10);
+        MenuTitle uut = new MenuTitle("Test");
         Output result = uut.render(colors);
+        String dividerText = (colors.length > 1 ? "[red]" : "[green]") + "*".repeat(10);
 
-        assertEquals(1, result.getOutput().size());
-        assertEquals(expected, result.getOutput().get(0));
+        assertEquals(4, result.getOutput().size());
+        assertEquals("", result.getOutput().get(0));
+        assertEquals(dividerText, result.getOutput().get(1));
+        assertEquals(expected, result.getOutput().get(2));
+        assertEquals(dividerText, result.getOutput().get(3));
     }
 
     private static Stream<Arguments> testRender() {
         return Stream.of(
-            Arguments.of(new Color[] {BLUE, RED}, "[red]**********"),
-            Arguments.of(new Color[] {BLUE}, "[green]**********"),
-            Arguments.of(new Color[] {}, "[green]**********")
+            Arguments.of(new Color[] {BLUE, RED}, "[red]* [blue] Test [red] *"),
+            Arguments.of(new Color[] {BLUE}, "[green]* [blue] Test [green] *"),
+            Arguments.of(new Color[] {}, "[green]* [yellow] Test [green] *")
         );
     }
 }
