+++ agonyforge-mud-demo/src/main/java/com/agonyforge/mud/demo/cli/question/CharacterEffortQuestion.java
@@ -7,10 +7,10 @@ import com.agonyforge.mud.core.web.model.Input;
 import com.agonyforge.mud.core.web.model.Output;
 import com.agonyforge.mud.core.web.model.WebSocketContext;
 import com.agonyforge.mud.demo.cli.RepositoryBundle;
-import com.agonyforge.mud.demo.cli.menu.MenuItem;
-import com.agonyforge.mud.demo.cli.menu.MenuPane;
-import com.agonyforge.mud.demo.cli.menu.MenuPrompt;
-import com.agonyforge.mud.demo.cli.menu.MenuTitle;
+import com.agonyforge.mud.core.cli.menu.impl.MenuItem;
+import com.agonyforge.mud.core.cli.menu.impl.MenuPane;
+import com.agonyforge.mud.core.cli.menu.impl.MenuPrompt;
+import com.agonyforge.mud.core.cli.menu.impl.MenuTitle;
 import com.agonyforge.mud.demo.model.constant.Effort;
 import com.agonyforge.mud.demo.model.impl.MudCharacter;
 import org.slf4j.Logger;
